+++ src/directives/formly-form.test.js
@@ -1,7 +1,11 @@
+/* eslint no-shadow:0 */
+/* eslint no-console:0 */
+/* eslint max-len:0 */
+/* eslint max-nested-callbacks:0 */
 import {expect} from 'chai';
 import testUtils from '../test.utils.js';
 import angular from 'angular-fix';
-import sinon from 'sinon';
+import _ from 'lodash';
 
 const {getNewField, input, basicForm} = testUtils;
 
@@ -63,15 +67,15 @@ describe('formly-form', () => {
     expect(scope.theForm).to.eq(isolateScope.theFormlyForm);
   });
 
-  it(`should warn if there's no FormController to be assigned`, () => {
-    const originalWarn = console.warn;
-    console.warn = sinon.spy();
+  it(`should warn if there's no FormController to be assigned`, inject(($log) => {
     compileAndDigest(`
       <formly-form model="model" fields="fields" form="theForm" id="my-formly-form" root-el="div"></formly-form>
     `);
-    expect(console.warn).to.have.been.called;
-    console.warn = originalWarn;
-  });
+    const log = $log.warn.logs[0];
+    expect($log.warn.logs).to.have.length(1);
+    expect(log[0]).to.equal('Formly Warning:');
+    expect(log[1]).to.equal('Your formly-form does not have a `form` property. Many functions of the form (like validation) may not work');
+  }));
 
   it(`should put the formControl on the field's scope when using a different form root element`, () => {
     scope.fields = [getNewField()];
@@ -160,7 +164,7 @@ describe('formly-form', () => {
           className: 'bar',
           fieldGroup: [
             {type: 'input', key: key++},
-            {type: 'input', key: key++}
+            {type: 'input', key: key++, defaultValue: 'bar'}
           ]
         },
         {type: 'input', key: key++},
@@ -240,6 +244,13 @@ describe('formly-form', () => {
         scope.model[field.key] = 'bar';
         scope.options.resetModel();
         expect(scope.model[field.key]).to.eq('foo');
+      });
+
+      it(`should allow you to call the child's initField from the parent`, () => {
+        const field = scope.fields[0].fieldGroup[1];
+        compileAndDigest(formWithOptions);
+        scope.model[field.key] = undefined;
+        scope.options.initFields();
       });
 
       it(`should have the same formState`, () => {
@@ -535,58 +546,144 @@ describe('formly-form', () => {
       }
     });
 
-    describe(`options`, () => {
-      describe(`updateInitialValue`, () => {
+    describe(`updateInitialValue`, () => {
 
-        it(`should update the initial value of the fields`, () => {
-          compileAndDigest(template);
-          const field = scope.fields[0];
-          expect(field.initialValue).to.equal('myFoo');
-          scope.model.foo = 'otherValue';
-          scope.options.updateInitialValue();
-          expect(field.initialValue).to.equal('otherValue');
-        });
-
-        it(`should reset to the updated initial value`, () => {
-          compileAndDigest(template);
-          const field = scope.fields[0];
-          scope.model.foo = 'otherValue';
-          scope.options.updateInitialValue();
-          scope.model.foo = 'otherValueAgain';
-          scope.options.resetModel();
-          expect(field.initialValue).to.equal('otherValue');
-          expect(scope.model.foo).to.equal('otherValue');
-        });
+      it(`should update the initial value of the fields`, () => {
+        compileAndDigest(template);
+        const field = scope.fields[0];
+        expect(field.initialValue).to.equal('myFoo');
+        scope.model.foo = 'otherValue';
+        scope.options.updateInitialValue();
+        expect(field.initialValue).to.equal('otherValue');
       });
 
-      describe(`removeChromeAutoComplete`, () => {
-        it(`should not have a hidden input when nothing is specified`, () => {
-          const el = compileAndDigest(template);
-          const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
-          expect(autoCompleteFixEl).to.be.null;
+      it(`should reset to the updated initial value`, () => {
+        compileAndDigest(template);
+        const field = scope.fields[0];
+        scope.model.foo = 'otherValue';
+        scope.options.updateInitialValue();
+        scope.model.foo = 'otherValueAgain';
+        scope.options.resetModel();
+        expect(field.initialValue).to.equal('otherValue');
+        expect(scope.model.foo).to.equal('otherValue');
+      });
+    });
+
+    describe(`removeChromeAutoComplete`, () => {
+      it(`should not have a hidden input when nothing is specified`, () => {
+        const el = compileAndDigest(template);
+        const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
+        expect(autoCompleteFixEl).to.be.null;
+      });
+
+      it(`should add a hidden input when specified as true`, () => {
+        scope.options.removeChromeAutoComplete = true;
+        const el = compileAndDigest(template);
+        const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
+        expect(autoCompleteFixEl).to.exist;
+      });
+
+      it(`should override the 'true' global configuration`, inject((formlyConfig) => {
+        formlyConfig.extras.removeChromeAutoComplete = true;
+        scope.options.removeChromeAutoComplete = false;
+        const el = compileAndDigest(template);
+        const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
+        expect(autoCompleteFixEl).to.be.null;
+      }));
+
+      it(`should be added regardless of the option if the global config is set`, inject((formlyConfig) => {
+        formlyConfig.extras.removeChromeAutoComplete = true;
+        const el = compileAndDigest(template);
+        const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
+        expect(autoCompleteFixEl).to.exist;
+      }));
+    });
+
+    describe(`useOneTimeBindings`, () => {
+      it(`should use one time bindings if globally specified`, inject((formlyConfig) => {
+        formlyConfig.extras.field = {
+          'useOneTimeBindings': {
+            'form-options': true
+          }
+        };
+        const el = compileAndDigest(template);
+        const fieldsEl = el[0].querySelector('[form-options="::options"]');
+        expect(fieldsEl).to.exist;
+      }));
+
+      it(`should use one time bindings if locally specified`, inject((formlyConfig) => {
+        const el = compileAndDigest(
+          '<formly-form form="vm.myForm" model="model" fields="fields" use-one-time-bindings="{ form: true }"></formly-form>'
+        );
+        const fieldsEl = el[0].querySelector('[form="::theFormlyForm"]');
+        expect(fieldsEl).to.exist;
+      }));
+    });
+
+    describe(`fieldTransform`, () => {
+      beforeEach(() => {
+        formlyConfig.extras.fieldTransform = fieldTransform;
+      });
+
+      it(`should throw an error if something is passed in and nothing is returned`, () => {
+        scope.fields = [getNewField()];
+        scope.options.fieldTransform = function() {
+          // I return nothing...
+        };
+        expect(() => compileAndDigest()).to.throw(/^Formly Error: fieldTransform must return an array of fields/);
+      });
+
+      it(`should allow you to transform field configuration`, () => {
+        scope.options.fieldTransform = fieldTransform;
+        const spy = sinon.spy(scope.options, 'fieldTransform');
+        doExpectations(spy);
+      });
+
+      it(`should use formlyConfig.extras.fieldTransform when not specified on options`, () => {
+        const spy = sinon.spy(formlyConfig.extras, 'fieldTransform');
+        doExpectations(spy);
+      });
+
+      function doExpectations(spy) {
+        const originalFields = [{
+          key: 'keyProp',
+          template: '<hr />',
+          customThing: 'foo',
+          otherCustomThing: {
+            whatever: '|-o-|'
+          }
+        }];
+        scope.fields = originalFields;
+        compileAndDigest();
+        expect(spy).to.have.been.calledWith(originalFields, scope.model, scope.options, scope.form);
+        const field = scope.fields[0];
+
+        expect(field).to.not.have.property('customThing');
+        expect(field).to.not.have.property('otherCustomThing');
+        expect(field).to.have.deep.property('data.customThing');
+        expect(field).to.have.deep.property('data.otherCustomThing');
+      }
+
+      function fieldTransform(fields) {
+        const extraKeys = ['customThing', 'otherCustomThing'];
+        return _.map(fields, field => {
+          var newField = {data: {}};
+          _.each(field, (val, name) => {
+            if (_.contains(extraKeys, name)) {
+              newField.data[name] = val;
+            } else {
+              newField[name] = val;
+            }
+          });
+          return newField;
         });
+      }
+    });
 
-        it(`should add a hidden input when specified as true`, () => {
-          scope.options.removeChromeAutoComplete = true;
-          const el = compileAndDigest(template);
-          const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
-          expect(autoCompleteFixEl).to.exist;
-        });
-
-        it(`should override the 'true' global configuration`, inject((formlyConfig) => {
-          formlyConfig.extras.removeChromeAutoComplete = true;
-          scope.options.removeChromeAutoComplete = false;
-          const el = compileAndDigest(template);
-          const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
-          expect(autoCompleteFixEl).to.be.null;
-        }));
-
-        it(`should be added regardless of the option if the global config is set`, inject((formlyConfig) => {
-          formlyConfig.extras.removeChromeAutoComplete = true;
-          const el = compileAndDigest(template);
-          const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]');
-          expect(autoCompleteFixEl).to.exist;
-        }));
+    describe(`data`, () => {
+      it(`should allow you to put whatever you want in data`, () => {
+        scope.options.data = {foo: 'bar'};
+        expect(compileAndDigest).to.not.throw();
       });
     });
   });